Changed precision of vectors in Medium to double.

+++ include/Linear.h
@@ -8,11 +8,11 @@ public:
     Linear(glm::vec3 _direction, double _slope, double _init);
     virtual ~Linear();
 
-    double f(glm::vec3 p);
-    glm::vec3 gradient(glm::vec3 p);
+    double f(glm::dvec3 p);
+    glm::dvec3 gradient(glm::dvec3 p);
 
 protected:
-    glm::vec3 direction;
+    glm::dvec3 direction;
     double slope;
     double init;
 };
