+++ include/Medium.h
@@ -8,6 +8,6 @@ public:
     Medium();
     virtual ~Medium();
 
-    virtual double f(glm::vec3 p) = 0;
-    virtual glm::vec3 gradient(glm::vec3 p) = 0;
+    virtual double f(glm::dvec3 p) = 0;
+    virtual glm::dvec3 gradient(glm::dvec3 p) = 0;
 };
